Oscilloscopeout
Had to include map and print header file in Oscilloscope.cpp
