+++ src/PqExtra.cpp
@@ -20,31 +20,6 @@
 
 #include "PqExtra.h"
 
-OscilloscopeOut::OscilloscopeOut(float minValue, float maxValue, uint8_t precision)
-  : _value(0), _minValue(minValue), _maxValue(maxValue), _precision(precision) {}
-
-float OscilloscopeOut::put(float value) {
-  // Copy value.
-  _value = value;
-
-  // Convert to bin.
-  float mapped = mapTo01(_value, _minValue, _maxValue);
-  int bin = round( mapped * _precision );
-  bin = constrain(bin, 0, _precision-1);
-
-  // Print.
-  print(_minValue, 2);
-  print(" |");
-  for (int i=0; i<_precision; i++)
-    print(i == bin ? '*' : ' ');
-  print("| ");
-  print(_maxValue, 2);
-  println();
-
-  // Return it.
-  return _value;
-}
-
 Smoother::Smoother(float factor)
   : PqPutter(),
     MovingAverage(factor) {
